pyproject.toml
[project]
name = "qa-navigator"
version = "0.1.0"
description = "AI-powered Visual QA Testing Agent - exhaustive UI testing via Gemini + ADK"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"

dependencies = [
    "google-adk>=1.25.0",
    "google-genai>=1.63.0",
    "playwright>=1.50.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
qa-navigator = "qa_navigator.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qa_navigator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
